Derive or Assign Arm Names

This function checks if arm_names is provided. If arm_names is missing, it attempts to derive names from mu_list. If mu_list does not contain names, it assigns default names ("A1", "A2", etc.) to each arm. Informational messages are displayed if verbose is set to TRUE.

derive_arm_names(arm_names, mu_list, verbose = FALSE)

Arguments

arm_names

Optional vector of arm names.

mu_list

Named list of means per treatment arm, from which arm names may be derived.

verbose

Logical, if TRUE, displays messages about the derivation process.

Value

A vector of arm names.

Author

Thomas Debray tdebray@fromdatatowisdom.com