Simulate a Parallel Design and Compute Ratio of Means (ROM)

Simulates a parallel-group design and calculates the p-value for the ratio of means (ROM) using an equivalence test.

test_par_rom(
  n,
  muT,
  muR,
  SigmaT,
  SigmaR,
  lequi_tol,
  uequi_tol,
  alpha,
  dropout,
  typey,
  adseq,
  k,
  arm_seedT,
  arm_seedR,
  TART,
  TARR,
  vareq
)

Arguments

n

integer number of subjects per arm

muT

vector mean of endpoints on treatment arm

muR

vector mean of endpoints on reference arm

SigmaT

matrix covar-variance matrix on treatment arm across endpoints

SigmaR

matrix covar-variance matrix on reference arm across endpoints

lequi_tol

vector lower equivalence tolerance band across endpoints

uequi_tol

vector upper equivalence tolerance band across endpoints

alpha

vector alpha value across endpoints

dropout

vector of size 2 with dropout proportion per arm (T,R)

typey

vector with positions of primary endpoints

adseq

boolean is used a sequential adjustment?

k

integer minimum number of equivalent endpoints

arm_seedT

integer seed for the simulation on treatment arm

arm_seedR

integer seed for the simulation on reference arm

TART

double treatment allocation rate for the treatment arm

TARR

double treatment allocation rate for the reference arm

vareq

Boolean. If TRUE, assumes equal variance between arms and applies Schuirmann’s two one-sided tests (TOST) for equivalence using a pooled variance.

Value

A numeric matrix containing the following simulated statistics:

ptost: p-values for equivalence testing.

mu0, mu1: Mean values for the reference and treatment groups.

sd0, sd1: Standard deviations for the reference and treatment groups.

totaly: Indicator (1/0) for equivalence success.

Details

The function simulates a parallel-group study design and applies an equivalence test using the ratio of means (ROM) approach.

Adjusts for dropout rates (dropout) and treatment allocation rates (TART, TARR).

Computes test statistics and determines whether the number of significant endpoints meets k.

Uses sequential testing if adseq = TRUE.

If vareq = TRUE, assumes equal variance between treatment and reference groups, applying Schuirmann’s two one-sided tests (TOST) for equivalence.